+++ Sources/KeyboardKit/Actions/KeyboardAction+Button.swift
@@ -32,7 +32,7 @@ public extension KeyboardAction {
         case .keyboardType(let type): type.standardButtonText(for: context)
         case .nextLocale: context.locale.languageCode?.uppercased()
         case .primary(let type): type.standardButtonText(for: context.locale)
-        case .space: KKL10n.space.text(for: context)
+        case .space: KKL10n.space.text(for: context.locale)
         case .text(let char): standardButtonText(for: char)
         default: nil
         }